# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Zerodrift Agent API",
    "version": "1.0.0",
    "description": "OpenAPI reference for upload, audit session, workflow, and session group endpoints."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current deployment"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Opaque platform token",
        "description": "Use the PLATFORM_TOKEN value as an Authorization: Bearer credential."
      }
    },
    "schemas": {
...
      "DeleteSessionGroupResponse": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean",
            "enum": [
              true
            ],
            "description": "Whether the Session Group was deleted."
          },
          "deletedSessionIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "IDs of Sessions deleted with the Group."
          },
          "groupId": {
            "type": "string",
            "description": "Deleted logical Session Group ID."
          }
        },
        "required": [
          "deleted",
          "deletedSessionIds",
          "groupId"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message."
          }
        },
        "required": [
          "error"
        ]
      },
...
      "SessionGroupCreateBody": {
        "type": "object",
        "properties": {
          "agentIds": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "description": "Agent Runtime IDs shared by every Session in the Group."
          },
          "groupName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120,
            "description": "Logical Session Group name; duplicate names are allowed."
          },
          "metadata": {
            "allOf": [
              {
                "$ref": "#/components/schemas/JsonObjectSchema"
              }
            ],
            "description": "Optional metadata applied to every created Session."
          },
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionGroupMemberCreateBody"
            },
            "minItems": 1,
            "maxItems": 500,
            "description": "Project-specific Session inputs, in the requested order."
          },
          "source": {
            "type": "string",
            "description": "Optional source label applied to every created Session."
          },
          "workflowId": {
            "type": "string",
            "minLength": 1,
            "description": "Workflow shared by every Session in the Group."
          }
        },
        "required": [
          "agentIds",
          "groupName",
          "sessions",
          "workflowId"
        ]
      },
...
      "SessionGroupCreateResponse": {
        "type": "object",
        "properties": {
          "created": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionGroupCreatedMember"
            },
            "description": "Successfully created Sessions, in input order."
          },
          "failed": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionGroupFailedMember"
            },
            "description": "Sessions that failed to create, in input order."
          },
          "groupId": {
            "type": "string",
            "description": "Created logical Session Group ID."
          },
          "groupName": {
            "type": "string",
            "description": "Stored Session Group name."
          }
        },
        "required": [
          "created",
          "failed",
          "groupId",
          "groupName"
        ]
      },
      "SessionGroupDetailsResponse": {
        "type": "object",
        "properties": {
          "createdAt": {
            "allOf": [
              {
                "$ref": "#/components/schemas/IsoDateTimeSchema"
              }
            ],
            "description": "When the logical Session Group was created."
          },
          "groupId": {
            "type": "string",
            "description": "Logical Session Group ID."
          },
          "groupName": {
            "type": "string",
            "description": "Stored Session Group name."
          },
          "reviewMinSeverity": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SeveritySchema"
              }
            ],
            "description": "Group-specific minimum severity included in Needs Review."
          },
          "sessions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "index": {
                  "type": "integer",
                  "minimum": 0,
                  "description": "Original input Session index."
                },
                "session": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/AuditSessionSnapshot"
                    }
                  ],
                  "description": "Current associated Session state."
                }
              },
              "required": [
                "index",
                "session"
              ]
            },
            "description": "Associated Sessions in original input order."
          }
        },
        "required": [
          "createdAt",
          "groupId",
          "groupName",
          "reviewMinSeverity",
          "sessions"
        ]
      },
...
      "SessionGroupReviewSettingsBody": {
        "type": "object",
        "properties": {
          "reviewMinSeverity": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SeveritySchema"
              }
            ],
            "description": "Minimum Finding severity included in Needs Review for this Group."
          }
        },
        "required": [
          "reviewMinSeverity"
        ]
      },
      "SessionGroupReviewSettingsResponse": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "string",
            "description": "Updated logical Session Group ID."
          },
          "reviewMinSeverity": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SeveritySchema"
              }
            ],
            "description": "Stored Group-specific Needs Review severity threshold."
          }
        },
        "required": [
          "groupId",
          "reviewMinSeverity"
        ]
      },
...
      "WorkflowDeleteResponse": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "workflowId": {
            "type": "string"
          }
        },
        "required": [
          "deleted",
          "workflowId"
        ]
      },
      "WorkflowDocumentInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120,
            "description": "Workflow name."
          },
          "category": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60,
            "description": "Optional Workflow category."
          },
          "defaultModel": {
            "type": "string",
            "minLength": 1,
            "description": "Default registered Agent ID for this Workflow."
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "description": "Optional Workflow description."
          },
          "stages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 120
                },
                "dependsOn": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "threadMode": {
                  "type": "string",
                  "enum": [
                    "new",
                    "fork",
                    "resume"
                  ]
                },
                "turns": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "prompt": {
                        "oneOf": [
                          {
                            "type": "string",
                            "minLength": 1
                          },
                          {
                            "type": "array",
                            "items": {
                              "type": "string"
                            },
                            "minItems": 1
                          }
                        ],
                        "description": "Instruction text or lines sent to the Stage Agent."
                      },
                      "goal": {
                        "type": "boolean"
                      },
                      "skills": {
                        "type": "object",
                        "properties": {
                          "names": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "prefixes": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        }
                      },
                      "mcp": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "outputSchema": {
                        "allOf": [
                          {
                            "$ref": "#/components/schemas/WorkflowOutputSchemaNameSchema"
                          }
                        ]
                      },
                      "beforeHooks": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "afterHooks": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "required": [
                      "prompt"
                    ]
                  },
                  "minItems": 1
                }
              },
              "required": [
                "name",
                "dependsOn",
                "threadMode",
                "turns"
              ]
            },
            "minItems": 1
          }
        },
        "required": [
          "name",
          "stages"
        ]
      },
...
      "WorkflowsResponse": {
        "type": "object",
        "properties": {
          "defaultWorkflowId": {
            "type": "string",
            "description": "Default Workflow ID."
          },
          "workflows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowDefinitionSummary"
            },
            "description": "Built-in and .data/workflow JSON Workflows."
          }
        },
        "required": [
          "defaultWorkflowId",
          "workflows"
        ]
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Authentication required"
                },
                "code": {
                  "type": "string",
                  "example": "401"
                }
              }
            }
          }
        }
      },
      "403": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Access denied"
                },
                "code": {
                  "type": "string",
                  "example": "403"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                },
                "code": {
                  "type": "string",
                  "example": "500"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/session-groups": {
      "post": {
        "operationId": "post-api-session-groups",
        "summary": "Create a logical Session Group",
        "description": "Creates a logical Session Group and independently queues each valid local or uploaded project as a normal Session.",
        "tags": [
          "Session Groups"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionGroupCreateBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created Session Group and per-Session results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionGroupCreateResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/api/session-groups/{groupId}": {
      "get": {
        "operationId": "get-api-session-groups-{groupId}",
        "summary": "Get a logical Session Group",
        "description": "Returns a logical Session Group and the current state of its associated Sessions.",
        "tags": [
          "Session Groups"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Session Group ID."
            },
            "description": "Session Group ID.",
            "example": "123"
          }
        ],
        "responses": {
          "200": {
            "description": "Session Group details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionGroupDetailsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "description": "Session Group not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "patch": {
        "operationId": "patch-api-session-groups-{groupId}",
        "summary": "Update the Needs Review severity threshold for a Session Group",
        "description": "",
        "tags": [
          "Session Groups"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Session Group ID."
            },
            "description": "Session Group ID.",
            "example": "123"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionGroupReviewSettingsBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated review settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionGroupReviewSettingsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "description": "Session Group not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "operationId": "delete-api-session-groups-{groupId}",
        "summary": "Delete a Session Group and its Sessions",
        "description": "Deletes the Group and every associated Session, including stored state and runtime artifacts. Active Groups cannot be deleted.",
        "tags": [
          "Session Groups"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Session Group ID."
            },
            "description": "Session Group ID.",
            "example": "123"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted Group and Session IDs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteSessionGroupResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "description": "Session Group not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/api/workflows": {
      "get": {
        "operationId": "get-api-workflows",
        "summary": "List audit workflows",
        "description": "Returns built-in and .data/workflow JSON workflows with the default workflow ID.",
        "tags": [
          "Workflows"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Saved audit workflows",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "operationId": "post-api-workflows",
        "summary": "Create a Workflow JSON document",
        "description": "Validates and saves one complete Workflow JSON document under .data/workflow.",
        "tags": [
          "Workflows"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowDocumentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created Workflow JSON document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDocumentResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/api/workflows/{workflowId}": {
      "get": {
        "operationId": "get-api-workflows-{workflowId}",
        "summary": "Get a Workflow JSON document",
        "description": "",
        "tags": [
          "Workflows"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "workflowId",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Stable Workflow ID."
            },
            "description": "Stable Workflow ID.",
            "example": "123"
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow JSON document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDocumentResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "put": {
        "operationId": "put-api-workflows-{workflowId}",
        "summary": "Save a Workflow JSON document",
        "description": "Validates and atomically replaces a custom Workflow JSON file.",
        "tags": [
          "Workflows"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "workflowId",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Stable Workflow ID."
            },
            "description": "Stable Workflow ID.",
            "example": "123"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowDocumentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved Workflow JSON document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDocumentResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "operationId": "delete-api-workflows-{workflowId}",
        "summary": "Delete a Workflow",
        "description": "Permanently deletes a custom Workflow JSON file. Built-in Workflows are read-only and existing Sessions retain their frozen snapshots.",
        "tags": [
          "Workflows"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "workflowId",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Stable Workflow ID."
            },
            "description": "Stable Workflow ID.",
            "example": "123"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted Workflow",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDeleteResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
...
}